# Conflicts:
#	mooncake-store/src/master_service.cpp

+++ mooncake-transfer-engine/include/transfer_metadata.h
@@ -59,7 +59,7 @@ class TransferMetadata {
    public:
     struct DeviceDesc {
         std::string name;
-        uint16_t lid;
+        uint32_t lid;
         std::string gid;
         std::string eid;  // for ub
 
@@ -185,7 +185,7 @@ class TransferMetadata {
     struct HandShakeDesc {
         std::string payload;  // opaque transport-specific handshake data
         std::string local_nic_path;
-        uint16_t local_lid = 0;
+        uint32_t local_lid = 0;
         std::string local_gid;
         std::string peer_nic_path;
 #ifdef USE_UB
@@ -260,6 +260,10 @@ class TransferMetadata {
     int addLocalMemoryBuffer(const BufferDesc &buffer_desc,
                              bool update_metadata);
 
+    // Removes the local descriptor entry before optional metadata publication.
+    // If publication fails, the local removal remains committed; a retry must
+    // publish the updated descriptor rather than remove the entry again.
+    // ERR_ADDRESS_NOT_REGISTERED means no matching local entry was present.
     int removeLocalMemoryBuffer(void *addr, bool update_metadata);
 
     int addLocalSegment(SegmentID segment_id, const std::string &segment_name,
